Add adjtime & adjtimex syscall
WARNING:
in linux, adjtime is not a syscall ! In fact glibc translates adjtime
calls into adjtimex calls. So here we do the same in DCE.

Problem is that when adding adjtimex to libos, we need to define the
timex structure. Including "linux/timex" triggers a lot of errors
because of redefinitions between libc and kernel headers. Thus for now,
I trick the system using a void parameter. A better interface should be
found.